// packages/components/src/ui/EmptyState/EmptyState.styles.d.ts
export declare const useStyles: () => {
  container: {
    display: 'flex';
    flexDirection: 'column';
    alignItems: 'center';
    justifyContent: 'center';
    gap: import('@alveole/theme').Spacing;
    paddingTop: import('@alveole/theme').Spacing;
    paddingBottom: import('@alveole/theme').Spacing;
    paddingLeft: import('@alveole/theme').Spacing;
    paddingRight: import('@alveole/theme').Spacing;
    textAlign: 'center';
    height: '100%' | undefined;
    width: '100%';
  };
  contenu: {
    display: 'flex';
    gap: import('@alveole/theme').Spacing;
    alignItems: 'center';
    textAlign: 'center';
    flex: number | undefined;
    justifyContent: 'center';
    width: '100%';
  };
  media: {
    display: 'flex';
    alignItems: 'center';
    justifyContent: 'center';
    color: '#002764';
    backgroundColor: '#F2F9FF';
    borderRadius: import('packages/theme/src/constants/Radius').Radius;
    padding: import('@alveole/theme').Spacing;
  };
  messageEtDescription: {
    display: 'flex';
    flexDirection: 'column';
    gap: import('@alveole/theme').Spacing;
    width: '100%';
    justifyContent: 'center';
  };
  title: {
    textAlign: 'center';
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 18 | 20;
    lineHeight: 24 | 28;
    letterSpacing: 0;
    color: '#151617';
  };
  description: {
    textAlign: 'center';
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 14;
    lineHeight: 20;
    letterSpacing: 0;
    color: '#5F6571';
  };
  footer: {
    width: '100%' | undefined;
    display: 'flex';
    flexDirection: 'column';
    gap: import('@alveole/theme').Spacing;
  };
};
//# sourceMappingURL=EmptyState.styles.d.ts.map